+++ apps/launcher/main.cpp
@@ -40,7 +40,7 @@ namespace {
 
 constexpr wchar_t window_class[] = L"Pal3MienLauncherWindow";
 constexpr wchar_t window_title[] = L"Palworld 3 Miền";
-constexpr std::string_view launcher_version = "1.0.35";
+constexpr std::string_view launcher_version = "1.0.36";
 constexpr std::string_view launcher_display_version = "1.0";
 constexpr std::string_view palverify_version = "1.0.16";
 constexpr std::string_view default_manifest_url =